requirements.txt
pycrypto
requests
argparse
getpass